+++ src/assembler.ts
@@ -10,21 +10,42 @@ import {
 import { toBase64 } from './utils/qrcodeGenerator';
 
 export function generatePixObject(elements: PixElements): PixObject {
-  switch (elements.type) {
-    case PixElementType.STATIC:
-      return {
-        ...elements,
-        toBRCode: () => createEmv(elements),
-        toImage: () => toBase64(createEmv(elements)),
-      } as PixStaticObject;
+  const emvCode = createEmv(elements);
+  if (elements.type === PixElementType.STATIC) {
+    const staticObject = {
+      ...elements,
+      toBRCode: () => emvCode,
+      toImage: () => toBase64(emvCode),
+    };
 
-    case PixElementType.DYNAMIC:
-      return {
-        ...elements,
-        toBRCode: () => createEmv(elements),
-        fetchPayload: ({ DPP, codMun }) =>
-          fetchPayload({ url: elements.url, DPP, codMun }),
-        toImage: () => toBase64(createEmv(elements)),
-      } as PixDynamicObject;
+    const throwIfError = () => ({
+      ...staticObject,
+      throwIfError: () => staticObject,
+    });
+
+    return {
+      ...staticObject,
+      throwIfError,
+    } as PixStaticObject;
   }
+  if (elements.type === PixElementType.DYNAMIC) {
+    const dynamicObject = {
+      ...elements,
+      toBRCode: () => emvCode,
+      fetchPayload: ({ DPP, codMun }) =>
+        fetchPayload({ url: elements.url, DPP, codMun }),
+      toImage: () => toBase64(emvCode),
+    };
+
+    const throwIfError = () => ({
+      ...dynamicObject,
+      throwIfError: () => dynamicObject,
+    });
+
+    return {
+      ...dynamicObject,
+      throwIfError,
+    } as PixDynamicObject;
+  }
+  throw new Error('Invalid Pix type');
 }